pyproject.toml
[tool.poetry]
name = "pydriller-go-ipfs"
version = "0.1.0"
description = ""
authors = ["Matthew Aylward <matthewtyleraylward@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyDriller = "^2.0"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
